app/requirements.txt
fastapi
boto3
requests
python-jose
pydantic[email,typing_extensions,dotenv]
starlette
python-dotenv